Commonise the fixtures in the bag indexer tests
